Cleaned up my Model package, resolved the errors within the DataHandler class.
I will now be working on connecting my SQLite database to the listView on my HomePage Fragment.

After the successful population of my ListView to the Database I will be testing the addCharacter, deleteCharacter, and updateCharacter methods.

Assuming all goes well on those fronts, I will then be connecting the SQLite database to a second listView on the Event Log fragment to actually track the changes made to the database.